Intel-specific workaround for LinkFlags tests
The Intel C compiler for Linux does not seem to reject any bad flags or
object files on its link lines.  Work around the problem by using a
preprocessor #error directive to ensure that BADFLAG appears in the
build output.  This does not really achieve the purpose of the tests but
it allows them to pass.