Add support for the emulation version of the cudart library.
In version 3.0 of the CUDA toolkit when building code for emulation, you need to link
against a new version of the cuda run time library called cudartemu.  This CL adds a check
for the new library and uses it when present and in emulation mode.  Note that this
library is not present in previous or subsequent versions of the CUDA toolkit.