11041: Improve FindCxxTest to use Python or Perl automatically; custom flags
Included patch by Simone Rossetto to check if either Python or Perl
are present in the system.  Whichever intepreter that is detected
is now used to run the test generator program.  If both interpreters
are detected, the CXXTEST_USE_PYTHON variable is obeyed.

Also added support for CXXTEST_TESTGEN_ARGS, for manually specifying
options to the CxxTest code generator.