Added CUDA 3.2 directory changes.  Disable emulation mode for CUDA 3.1+.
CUDA 3.2 on Windows systems changed the layout of the library paths.  This adds the extra
directories needed to locate the new files.

I also explicitly disable emulation mode for CUDA 3.1+.  This feature was deprecated in
3.0 and ultimately removed in 3.1.  The script errors out if CUDA_BUILD_EMULATION is
turned on.  I didn't want to ignore emulation mode (even with a warning - which most
people may not even see) and have users confused as to why it wasn't working.