ccmake: Fix search with '/'
Commit 7a18dd8e (Add searching of variables, 2003-03-07) added method
cmCursesMainForm::JumpToCacheEntry to search for cache entries whose
names match a given search string.  The method also had a useless
argument "int idx" probably left from earlier development iterations and
hard-coded in all calls to the value '-1'.  The method compared this
argument to the "NumberOfVisibleEntries" member which at the time was of
type "int" also.

Commit ff1f8d0b (Fix or cast more integer conversions in cmake,
2010-06-29) changed the type of "NumberOfVisibleEntries" to size_t to
fix other integer conversion warnings.  An unsigned type makes sense
given the purpose of the member.  However, this caused the '-1' signed
value to be converted to a large unsigned value in the above-mentioned
comparison, leading to incorrect behavior.

Fix the problem by removing the useless argument and the comparison.