CodeBlocks Generator: Do not omit files in the project file listing.
The CodeBlocks generator used to omit files that were in siblings directories.
Though that filters out a few files that could be argued aren't of interest
to the user, it also potentialy hides files that are relevant to the user.

Patch by Daniel Teske.

Alex

See: http://bugreports.qt.nokia.com/browse/QTCREATORBUG-2250 for an example.