No CMake.HTML test if xmllint has no --nonet.
In commit bb1df1ec, we temporarily ran an alternate test,
guaranteed to fail when the --help output of xmllint did
not contain --nonet and --path.

This commit simply eliminates the test altogether in
this condition rather than make an attempt (doomed to
fail) to pull down the dtd over the internet.

On date=2010-11-04, the CMake dashboard results showed that
the test failed on the following CMake dashboard machines:

  dash8.kitware
  dash8.kitwarein.com
  dashsun1
  dashsun1.kitware
  ferrari

This is a very small subset of the dashboard machines, and
we have enough proof from enough other machines that the test
passes with xmllint versions new enough to have the --nonet
support.

Therefore, eliminate the CMake.HTML test on machines with old
versions of xmllint. To run the test, make sure you run it
on a machine with a new enough xmllint.