Cygwin: Fix release script libncurses search patterns (#10766)
With the switch to upstream ncurses "ABI 6", Cygwin's ncurses has YA ABI
bump and is now libncurses10.  However, the regex used to determine in
Utilities/Release/Cygwin/CMakeLists.txt which libncurses is being used
does not handle multiple-digit ABIs.

libncurses8 was the first version to be built with libtool and therefore
contains a hyphen (cygncurses-8.dll).  It was first introduced in 2004,
so it should be sufficiently old to rely on.  Furthermore, libncurses7
has a serious flaw in that it completely breaks if rebased.

Therefore the easiest solution is to only look at the hyphened versions
and change the regex accordingly.